+++ packages/datadog_flutter_plugin/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * Deprecation - `DdSdkConfiguration.customEndpoint` has been deprecated in favor of `DdSdkConfiguration.customLogsEndpoint` and `RumConfiguration.customEndpoint`.
 * Added `DdSdkConfiguration.version` configuration option for specifying a custom application version.
-* Fix `null` values in attributes not being correctly encoded on iOS
+* Fix `null` values in attributes not being correctly encoded on iOS.
+* Add `flavor` as a configuration parameter.
 
 ## 1.0.0-rc.3